+++ server/src/providers/not-supported.ts
@@ -1,0 +1,6 @@
+export class NotSupportedError extends Error {
+  constructor(method: string) {
+    super(`NotSupported: ${method}`);
+    this.name = 'NotSupportedError';
+  }
+}